CTS/requirements.txt
asgiref==3.4.1
coloredlogs==15.0.1
Django==4.0
django-seed==0.3.1
djangorestframework==3.13.1
execute==1.2
executor==23.2
Faker==10.0.0
fasteners==0.16.3
humanfriendly==10.0
property-manager==3.0
psycopg2==2.9.2
pyreadline3==3.3
python-dateutil==2.8.2
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
text-unidecode==1.3
toposort==1.7
tzdata==2021.5
verboselogs==1.7
